Move CRS tests to separate test file
